# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FDDT — Fraud Document Detection Tool API",
    "description": "FDDT (Fraud Document Detection Tool) backend. Environment: local.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/cases": {
      "post": {
        "tags": [
          "cases"
        ],
        "summary": "Create a case",
        "description": "Opens an empty case (status `submitted`) owned by the caller, in the caller's company; any company role may create one (platform admins may not — 403). The case number is `CASE-` plus the first 8 hex characters of the case UUID. Attach documents afterwards with POST /cases/{case_id}/documents. Writes a `case_created` audit event.",
        "operationId": "create_case_cases_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired bearer token, or the user is inactive."
          },
          "403": {
            "description": "Platform admins cannot create cases."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "cases"
        ],
        "summary": "List cases (the case queue)",
        "description": "Only the caller's company. Reviewers see every case of their company; a `user` sees only the cases they submitted. A platform admin must pass `company_id` (one company at a time, audited) — there is no cross-company case list. For `reviewer_l2` and platform admins, open cases escalated to L2 sort first; otherwise newest first. Optional filters: `status`, `case_type`, `assigned_tier`, and `actionable=true` (only cases the caller may act on — for a `reviewer_l1` that excludes L2 cases). Each item carries `can_act`. For submitters `risk_tier` is null and `flag` is a neutral 'In review' marker, so scoring signals are never disclosed to them.",
        "operationId": "list_cases_cases_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CaseStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "case_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CaseType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Case Type"
            }
          },
          {
            "name": "assigned_tier",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CaseTier"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assigned Tier"
            }
          },
          {
            "name": "actionable",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Only cases the caller may approve/reject/escalate.",
              "default": false,
              "title": "Actionable"
            },
            "description": "Only cases the caller may approve/reject/escalate."
          },
          {
            "name": "company_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Platform admins only (required for them): the company to act on. Company users are always confined to their own company.",
              "title": "Company Id"
            },
            "description": "Platform admins only (required for them): the company to act on. Company users are always confined to their own company."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CaseListItem"
                  },
                  "title": "Response List Cases Cases Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired bearer token, or the user is inactive."
          },
          "400": {
            "description": "A platform admin did not pass `company_id`."
          },
          "404": {
            "description": "`company_id` names an unknown company (or, for a company user, not theirs)."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/cases/{case_id}/documents/{document_id}/signature-references": {
      "post": {
        "tags": [
          "signatures"
        ],
        "summary": "Create a signature reference",
        "description": "Create a signature reference from a reviewer-drawn bounding box.\n\nCrops the signature region from the document's PDF, uploads it to\nBlob Storage, inserts the signature_references row, then immediately\nenqueues the in-case comparison Celery task for this reference.\n\n`person_name` is required and must be typed by the reviewer — the\nsystem does not attempt to auto-extract it (System Specification §2.3).",
        "operationId": "create_signature_reference_cases__case_id__documents__document_id__signature_references_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignatureReferenceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignatureReferenceResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired bearer token, or the user is inactive."
          },
          "403": {
            "description": "Platform admins have read-only access."
          },
          "404": {
            "description": "No such case (or, for a `user`, not one they submitted), or the document is not in this case."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}/signature-references": {
      "get": {
        "tags": [
          "signatures"
        ],
        "summary": "List a case's signature references",
        "description": "All signature references created for this case. Used by the upload\nscreen to list what references were set during this session.",
        "operationId": "list_signature_references_cases__case_id__signature_references_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SignatureReferenceResponse"
                  },
                  "title": "Response List Signature References Cases  Case Id  Signature References Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired bearer token, or the user is inactive."
          },
          "404": {
            "description": "No such case (or, for a `user`, not one they submitted)."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}/signature-matches": {
      "get": {
        "tags": [
          "signatures"
        ],
        "summary": "List a case's signature comparison results",
        "description": "All in-case signature match results for this case, enriched with the\nreference person_name. Filtered to comparison_scope='in_case' only —\nlibrary-scope matches are not generated yet (System Specification §2.3).",
        "operationId": "list_signature_matches_cases__case_id__signature_matches_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SignatureMatchResponse"
                  },
                  "title": "Response List Signature Matches Cases  Case Id  Signature Matches Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired bearer token, or the user is inactive."
          },
          "404": {
            "description": "No such case (or, for a `user`, not one they submitted)."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CaseCreateRequest": {
        "properties": {
          "case_type": {
            "$ref": "#/components/schemas/CaseType",
            "description": "Case-level category chosen at submission. Independent of each document's own `document_type`, which the classification step fills in later."
          }
        },
        "type": "object",
        "required": [
          "case_type"
        ],
        "title": "CaseCreateRequest"
      },
...
      "CaseFlagSchema": {
        "properties": {
          "flag": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "pending"
            ],
            "title": "Flag"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "flag",
          "label",
          "description"
        ],
        "title": "CaseFlagSchema",
        "description": "Wire shape of app/services/case_flag_service.py's CaseFlag: the\ncase's real risk tier (low/medium/high), or \"pending\" until scored."
      },
      "CaseListItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "case_number": {
            "type": "string",
            "title": "Case Number"
          },
          "case_type": {
            "$ref": "#/components/schemas/CaseType"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "assigned_tier": {
            "$ref": "#/components/schemas/CaseTier"
          },
          "risk_tier": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskTier"
              },
              {
                "type": "null"
              }
            ]
          },
          "submitted_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "document_count": {
            "type": "integer",
            "title": "Document Count"
          },
          "flag": {
            "$ref": "#/components/schemas/CaseFlagSchema"
          },
          "can_act": {
            "type": "boolean",
            "title": "Can Act",
            "description": "Whether the caller may approve/reject/escalate this case given its tier. False for submitters, and for a `reviewer_l1` on an L2-assigned case (read-only). Does not reflect status — a decided case is still `can_act` but every action is a 409."
          }
        },
        "type": "object",
        "required": [
          "id",
          "case_number",
          "case_type",
          "status",
          "assigned_tier",
          "risk_tier",
          "created_at",
          "document_count",
          "flag",
          "can_act"
        ],
        "title": "CaseListItem"
      },
...
      "CaseResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "case_number": {
            "type": "string",
            "title": "Case Number"
          },
          "case_type": {
            "$ref": "#/components/schemas/CaseType"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus"
          },
          "assigned_tier": {
            "$ref": "#/components/schemas/CaseTier"
          },
          "risk_tier": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskTier"
              },
              {
                "type": "null"
              }
            ]
          },
          "submitted_by_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Submitted By User Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "case_number",
          "case_type",
          "status",
          "assigned_tier",
          "risk_tier",
          "submitted_by_user_id",
          "created_at"
        ],
        "title": "CaseResponse"
      },
      "CaseStatus": {
        "type": "string",
        "enum": [
          "submitted",
          "under_automated_review",
          "pending_manual_review",
          "auto_approved",
          "escalated",
          "under_investigation",
          "approved",
          "rejected",
          "closed"
        ],
        "title": "CaseStatus"
      },
      "CaseTier": {
        "type": "string",
        "enum": [
          "l1",
          "l2"
        ],
        "title": "CaseTier",
        "description": "Which reviewer tier owns the case — orthogonal to `status`. Every case\nstarts at `l1`; escalating it moves it to `l2` (one-directional, there is\nno de-escalation). On an `l2` case a `reviewer_l1` keeps read access but\nevery action endpoint returns 403; `reviewer_l2` and `admin` can act on\ncases of either tier."
      },
      "CaseType": {
        "type": "string",
        "enum": [
          "school_document",
          "vendor_invoice",
          "commercial_invoice",
          "procurement_documentation",
          "quotation",
          "travel_reimbursement",
          "other"
        ],
        "title": "CaseType",
        "description": "Case-level category, selected at submission time (System Specification section\n3.6's document-type coverage list). This is separate from a document's\nown `document_type`, which is filled in later by the classification\nstep — a case can hold documents of more than one type (e.g. an\ninvoice plus its approval voucher)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RiskTier": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "RiskTier"
      },
...
      "SignatureMatchResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "case_id": {
            "type": "string",
            "format": "uuid",
            "title": "Case Id"
          },
          "signature_reference_id": {
            "type": "string",
            "format": "uuid",
            "title": "Signature Reference Id"
          },
          "reference_person_name": {
            "type": "string",
            "title": "Reference Person Name"
          },
          "comparison_scope": {
            "type": "string",
            "title": "Comparison Scope"
          },
          "result": {
            "type": "string",
            "title": "Result"
          },
          "result_label": {
            "type": "string",
            "title": "Result Label"
          },
          "reasoning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning"
          },
          "compared_at": {
            "type": "string",
            "format": "date-time",
            "title": "Compared At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "case_id",
          "signature_reference_id",
          "reference_person_name",
          "comparison_scope",
          "result",
          "result_label",
          "reasoning",
          "compared_at"
        ],
        "title": "SignatureMatchResponse",
        "description": "One signature_matches row returned by the API, enriched with the\nreference's person_name so the frontend can show \"compared against [name]\"\nwithout a separate reference lookup."
      },
      "SignatureReferenceCreate": {
        "properties": {
          "person_name": {
            "type": "string",
            "minLength": 1,
            "title": "Person Name",
            "description": "Name of the person whose signature this is, typed by the reviewer. Required — the system does not attempt to auto-extract this field."
          },
          "bounding_box": {
            "$ref": "#/components/schemas/BoundingBoxSchema"
          },
          "is_library": {
            "type": "boolean",
            "title": "Is Library",
            "description": "Defaults to True: every reference is saved to the permanent signature library. False is still accepted (case-only reference). Either way, at this phase a reference is only compared against other documents in its own case — no cross-case/library-wide comparison logic runs yet.",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "person_name",
          "bounding_box"
        ],
        "title": "SignatureReferenceCreate",
        "description": "Request body for POST /cases/{case_id}/documents/{document_id}/\nsignature-references.\n\n`person_name` is required and must be provided by the reviewer —\nthe system never attempts to auto-extract or pre-fill it (System Specification §2.3)."
      },
      "SignatureReferenceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "person_name": {
            "type": "string",
            "title": "Person Name"
          },
          "signature_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signature Image Url"
          },
          "bounding_box": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bounding Box"
          },
          "source_document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Source Document Id"
          },
          "source_case_id": {
            "type": "string",
            "format": "uuid",
            "title": "Source Case Id"
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          },
          "is_library": {
            "type": "boolean",
            "title": "Is Library"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "person_name",
          "signature_image_url",
          "bounding_box",
          "source_document_id",
          "source_case_id",
          "created_by",
          "is_library",
          "created_at"
        ],
        "title": "SignatureReferenceResponse",
        "description": "One signature_references row returned by the API."
      },
...
      "UserSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "full_name"
        ],
        "title": "UserSummary"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}